+++ app/build.gradle.kts
@@ -127,13 +127,14 @@ dependencies {
     implementation(projects.ivyCore)
     implementation(projects.ivyResources)
     implementation(projects.ivyNavigation)
+    implementation(projects.widgetBalance)
+    implementation(projects.widgetAddTransaction)
 
     implementation(libs.ivy.frp.temp)
     implementation(libs.bundles.kotlin)
     implementation(libs.bundles.ktor)
     implementation(libs.bundles.arrow)
     implementation(libs.bundles.compose)
-    implementation(libs.bundles.glance)
     implementation(libs.bundles.activity)
     implementation(libs.bundles.google)
     implementation(libs.bundles.firebase)
