+++ ivy-widget-base/src/main/java/com/ivy/widget/AddTransactionWidgetClick.kt
@@ -1,4 +1,4 @@
-package com.ivy.widgets
+package com.ivy.widget
 
 import android.app.PendingIntent
 import android.appwidget.AppWidgetManager
