+++ settings.gradle.kts
@@ -17,3 +17,5 @@ include(":ivy-core")
 include(":ivy-navigation")
 include(":ivy-resources")
 include(":ivy-widget-base")
+include(":widget-add-transaction")
+include(":widget-balance")
