+++ templates/__module/build.gradle.kts
@@ -9,4 +9,5 @@ android {
 dependencies {
     implementation(projects.ivyCore)
     implementation(projects.ivyResources)
+    implementation(projects.ivyDesign)
 }
