+++ widget-add-transaction/src/main/java/com/ivy/widget/transaction/AddTransactionWidgetCompact.kt
@@ -1,4 +1,4 @@
-package com.ivy.widgets
+package com.ivy.widget.transaction
 
 import android.appwidget.AppWidgetManager
 import android.appwidget.AppWidgetProvider
@@ -7,11 +7,11 @@ import android.content.Intent
 import android.widget.RemoteViews
 import com.ivy.wallet.R
 
-class AddTransactionWidget : AppWidgetProvider() {
+class AddTransactionWidgetCompact : AppWidgetProvider() {
 
     companion object {
         fun updateBroadcast(context: Context) {
-            WidgetBase.updateBroadcast(context, AddTransactionWidget::class.java)
+            WidgetBase.updateBroadcast(context, AddTransactionWidgetCompact::class.java)
         }
     }
 
@@ -36,17 +36,14 @@ class AddTransactionWidget : AppWidgetProvider() {
         appWidgetManager: AppWidgetManager,
         appWidgetId: Int
     ) {
-        val rv = RemoteViews(context.packageName, R.layout.widget_add_transaction)
+        val rv = RemoteViews(context.packageName, R.layout.widget_add_transaction_compact)
         val clickSetup = AddTransactionWidgetClick.Setup(context, rv, appWidgetId)
 
         clickSetup.clickListener(R.id.ivIncome, AddTransactionWidgetClick.ACTION_ADD_INCOME)
-        clickSetup.clickListener(R.id.tvIncome, AddTransactionWidgetClick.ACTION_ADD_INCOME)
 
         clickSetup.clickListener(R.id.ivExpense, AddTransactionWidgetClick.ACTION_ADD_EXPENSE)
-        clickSetup.clickListener(R.id.tvExpense, AddTransactionWidgetClick.ACTION_ADD_EXPENSE)
 
         clickSetup.clickListener(R.id.ivTransfer, AddTransactionWidgetClick.ACTION_ADD_TRANSFER)
-        clickSetup.clickListener(R.id.tvTransfer, AddTransactionWidgetClick.ACTION_ADD_TRANSFER)
 
         appWidgetManager.updateAppWidget(appWidgetId, rv)
     }
