+++ widget-add-transaction/src/main/res/xml/add_transaction_widget_compact_info.xml
