+++ widget-add-transaction/src/main/res/xml/add_transaction_widget_info.xml
