+++ widget-add-transaction/src/main/res/xml/wallet_balance_widget_info.xml
