+++ widget-balance/build.gradle.kts
@@ -3,10 +3,12 @@ plugins {
 }
 
 android {
-    namespace = "com.ivy.widget"
+    namespace = "com.ivy.widget.balance"
 }
 
 dependencies {
     implementation(projects.ivyCore)
     implementation(projects.ivyResources)
+    implementation(projects.ivyDesign)
+    implementation(projects.ivyWidgetBase)
 }
