+++ widget-balance/src/main/res/xml/wallet_balance_widget_info.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
+    android:resizeMode="horizontal|vertical"
+    android:previewImage="@drawable/preview_widget_wallet_balance"
+    android:minResizeWidth="250dp"
+    android:minWidth="250dp"
+    android:minHeight="110dp"
+    android:minResizeHeight="110dp"
+    android:widgetCategory="home_screen"
+/>
